+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ Ongoing
 - Added :class:`~.SanStream` for non-waste streams (e.g., gases).
 - Updated the ``add_OPEX`` attribute of :class:`~.SanUnit` and ``system_add_OPEX`` attribute of :class:`~.SimpleTEA` so that they take :class:`dict` as the default to allow display of multiple additional operating expenses.
 - Split the ``systems`` module into an individual package `EXPOsan <https://github.com/QSD-Group/exposan>`_.
+- Now using :class:`thermosteam.utils.Registry` to manage :class:`~.ImpactIndicator` and :class:`~.ImpactItem`.
 - Added `AppVeyor CI <https://ci.appveyor.com/project/yalinli2/qsdsan>`_.
 - Renamed the ``master`` branch to ``main``.